Add db/update.go for dynamic object updates
The thing-object doesn't really demonstrate its usefulness, but this
allows the object-creator to just store any object without worrying
about which parts are set or not.

For gondul, this is crucial since it allows partially updating an
existing object without zeroing out the preset values. E.g.: You can
update just the VlanID without having to manually merge it with
everything else.

Next up is a generic Get/Select, but that requires some extra thought.